+++ src/firmware/lib/blinkscript/src/parser/ast/expressions/array_expression.hpp
@@ -11,14 +11,14 @@ namespace SyncBlink
     class ArrayExpression : public AstNode
     {
     public:
-        explicit ArrayExpression(std::vector<std::unique_ptr<const AstNode>>&& arrayContent, int line);
+        explicit ArrayExpression(std::vector<std::unique_ptr<const AstNode>>&& arrayContent, uint16_t line);
 
         void accept(AstVisitor& visitor) const;
-        int getLine() const;
+        uint16_t getLine() const;
         const std::vector<std::unique_ptr<const AstNode>>& getArrayContent() const;
 
     private:
-        int _line;
+        uint16_t _line;
         std::vector<std::unique_ptr<const AstNode>> _arrayContent;
     };
 }
