+++ src/firmware/lib/blinkscript/src/parser/ast/expressions/for_expression.cpp
@@ -5,7 +5,7 @@ namespace SyncBlink
     ForExpression::ForExpression(std::unique_ptr<const AstNode> assignStatement,
                                  std::unique_ptr<const AstNode> conditionExpression,
                                  std::unique_ptr<const AstNode> incrementorStatement,
-                                 std::unique_ptr<const AstNode> loopBody, int line)
+                                 std::unique_ptr<const AstNode> loopBody, uint16_t line)
         : _assignStatement(std::move(assignStatement)), _conditionExpression(std::move(conditionExpression)),
           _incrementorStatement(std::move(incrementorStatement)), _loopBody(std::move(loopBody)), _line(line)
     {
@@ -31,7 +31,7 @@ namespace SyncBlink
     {
         return *_loopBody;
     }
-    int ForExpression::getLine() const
+    uint16_t ForExpression::getLine() const
     {
         return _line;
     }
