+++ src/firmware/lib/blinkscript/src/parser/ast/expressions/for_expression.hpp
@@ -13,17 +13,17 @@ namespace SyncBlink
         ForExpression(std::unique_ptr<const AstNode> assignStatement,
                       std::unique_ptr<const AstNode> conditionExpression,
                       std::unique_ptr<const AstNode> incrementorStatement,
-                      std::unique_ptr<const AstNode> loopBody, int line);
+                      std::unique_ptr<const AstNode> loopBody, uint16_t line);
 
         void accept(AstVisitor& visitor) const;
         const AstNode& getAssignStatement() const;
         const AstNode& getConditionExpression() const;
         const AstNode& getIncrementorStatement() const;
         const AstNode& getLoopBody() const;
-        int getLine() const;
+        uint16_t getLine() const;
 
     private:
-        int _line;
+        uint16_t _line;
         std::unique_ptr<const AstNode> _assignStatement;
         std::unique_ptr<const AstNode> _conditionExpression;
         std::unique_ptr<const AstNode> _incrementorStatement;
